Add tests for `WebhookTrigger` mutation.
